Move PokemonType to the Data layer
