+++ Sources/Data/PokemonType.swift
@@ -1,0 +1,22 @@
+import SwiftUI
+
+public enum PokemonType: String, CaseIterable {
+    case normal
+    case fire
+    case water
+    case grass
+    case electric
+    case ice
+    case fighting
+    case poision
+    case ground
+    case flying
+    case psychic
+    case bug
+    case rock
+    case ghost
+    case dark
+    case dragon
+    case steel
+    case fairy
+}
