+++ Tests/Data/PokemonTypeTests.swift
@@ -1,0 +1,7 @@
+import XCTest
+
+final class PokemonTypeTests: XCTestCase {
+    private var decoder: JSONDecoder
+    func test() {
+    }
+}
